requirements.txt
langchain==0.3.25
langchain-aws==0.2.14
langchain-community==0.3.24
langgraph==0.2.73
boto3==1.38.23
pypdf==5.4.0
pdfplumber==0.11.6
pytesseract==0.3.13
pdf2image==1.17.0
python-dotenv==1.1.0
